bot/requirements.txt
pyTelegramBotAPI
SQLAlchemy
python-dotenv
psycopg2-binary
